cluster: relocate centroid pipeline to experiments/datakit/cluster/v0/
Carves out a v0/ sub-package sibling to weborganizer/. Moves
assign / train / sample / summarize / exp_smoke / exp_full_clusters
and the run journal under v0/, updates the internal imports and
``python -m`` module paths, and patches the ClusterAssignAttrData
docstring reference in datakit_store.py to point at the new path.
No behavior change.
