+++ experiments/datakit/cluster/v0/exp_full_clusters.py
@@ -30,7 +30,7 @@ Submit::
     uv run iris --cluster=marin job run --no-wait --cpu=1 --memory=2G \\
         --extra=cpu \\
         --job-name "embed-clusters-full-$(date +%Y%m%d-%H%M%S)" \\
-        -- python -m experiments.datakit.cluster.exp_full_clusters
+        -- python -m experiments.datakit.cluster.v0.exp_full_clusters
 """
 
 import logging
@@ -47,10 +47,10 @@ from marin.execution.remote import remote  # noqa: E402
 from marin.execution.step_runner import StepRunner  # noqa: E402
 from marin.execution.step_spec import StepSpec  # noqa: E402
 
-from experiments.datakit.cluster.assign import AssignmentAttrData, assign_source  # noqa: E402
-from experiments.datakit.cluster.sample import sample_centroid_inputs  # noqa: E402
-from experiments.datakit.cluster.summarize import summarize_at_k  # noqa: E402
-from experiments.datakit.cluster.train import train_centroids  # noqa: E402
+from experiments.datakit.cluster.v0.assign import AssignmentAttrData, assign_source  # noqa: E402
+from experiments.datakit.cluster.v0.sample import sample_centroid_inputs  # noqa: E402
+from experiments.datakit.cluster.v0.summarize import summarize_at_k  # noqa: E402
+from experiments.datakit.cluster.v0.train import train_centroids  # noqa: E402
 from experiments.datakit.embeddings.luxical.pipeline import (  # noqa: E402
     LUXICAL_REPO,
     LUXICAL_WEIGHTS_FILE,
